test(evm): move block accumulator tests
