+++ crates/evm2/src/evm/state/mod.rs
@@ -1325,183 +1325,6 @@ mod tests {
         assert!(changes.storage.get(&address).is_some_and(|storage| storage.wipe));
     }
 
-    fn account_change(
-        address: Address,
-        original: Option<AccountInfo>,
-        current: Option<AccountInfo>,
-    ) -> StateChanges {
-        let mut changes = StateChanges::default();
-        changes.accounts.insert(address, Tracked::from_parts(original, current));
-        changes
-    }
-
-    fn storage_change(
-        address: Address,
-        key: Word,
-        original: Word,
-        current: Word,
-        wipe: bool,
-    ) -> StateChanges {
-        let mut changes = StateChanges::default();
-        changes.storage.insert(
-            address,
-            StorageChangeSet {
-                wipe,
-                slots: U256Map::from_iter([(key, Tracked::from_parts(original, current))]),
-                _non_exhaustive: (),
-            },
-        );
-        changes
-    }
-
-    fn storage_wipe(address: Address) -> StateChanges {
-        let mut changes = StateChanges::default();
-        changes.storage.insert(
-            address,
-            StorageChangeSet { wipe: true, slots: U256Map::default(), _non_exhaustive: () },
-        );
-        changes
-    }
-
-    fn without_code(mut info: AccountInfo) -> AccountInfo {
-        info.code = None;
-        info
-    }
-
-    #[test]
-    fn block_accumulator_collapses_create_then_delete() {
-        let address = Address::from([0x50; 20]);
-        let key = Word::from(1);
-        let created = AccountInfo::default().with_nonce(1);
-        let mut accumulator = BlockStateAccumulator::new();
-
-        let mut create = account_change(address, None, Some(created.clone()));
-        create.storage = storage_change(address, key, Word::ZERO, Word::from(7), true).storage;
-        create.visit(&mut accumulator).expect("block accumulator is infallible");
-
-        let mut delete = account_change(address, Some(created), None);
-        delete.storage = storage_wipe(address).storage;
-        delete.visit(&mut accumulator).expect("block accumulator is infallible");
-
-        assert!(accumulator.accounts_sorted().is_empty());
-        assert!(accumulator.storage_wipes_sorted().is_empty());
-        assert!(accumulator.storage_sorted().is_empty());
-    }
-
-    #[test]
-    fn block_accumulator_preserves_original_for_delete_then_recreate() {
-        let address = Address::from([0x51; 20]);
-        let key = Word::from(1);
-        let original = AccountInfo::default().with_balance(Word::from(3));
-        let recreated = AccountInfo::default().with_nonce(1);
-        let mut accumulator = BlockStateAccumulator::new();
-
-        let mut delete = account_change(address, Some(original.clone()), None);
-        delete.storage = storage_wipe(address).storage;
-        delete.visit(&mut accumulator).expect("block accumulator is infallible");
-
-        let mut create = account_change(address, None, Some(recreated.clone()));
-        create.storage = storage_change(address, key, Word::ZERO, Word::from(7), true).storage;
-        create.visit(&mut accumulator).expect("block accumulator is infallible");
-
-        let accounts = accumulator.accounts_sorted();
-        assert_eq!(accounts.len(), 1);
-        assert_eq!(accounts[0].1.original().as_ref(), Some(&without_code(original)));
-        assert_eq!(accounts[0].1.current().as_ref(), Some(&without_code(recreated)));
-        assert_eq!(accumulator.storage_wipes_sorted(), [address]);
-
-        let storage = accumulator.storage_sorted();
-        assert_eq!(storage.len(), 1);
-        assert_eq!(storage[0].0.key(), key);
-        assert_eq!(*storage[0].1.current(), Word::from(7));
-    }
-
-    #[test]
-    fn block_accumulator_keeps_nonzero_write_after_storage_wipe() {
-        let address = Address::from([0x52; 20]);
-        let key = Word::from(1);
-        let original = AccountInfo::default().with_balance(Word::from(3));
-        let mut accumulator = BlockStateAccumulator::new();
-
-        let mut wipe_and_restore = account_change(address, Some(original.clone()), Some(original));
-        wipe_and_restore.storage =
-            storage_change(address, key, Word::from(5), Word::from(5), true).storage;
-        wipe_and_restore.visit(&mut accumulator).expect("block accumulator is infallible");
-
-        assert!(accumulator.accounts_sorted().is_empty());
-        assert_eq!(accumulator.storage_wipes_sorted(), [address]);
-
-        let storage = accumulator.storage_sorted();
-        assert_eq!(storage.len(), 1);
-        assert_eq!(storage[0].0.key(), key);
-        assert_eq!(*storage[0].1.current(), Word::from(5));
-    }
-
-    #[test]
-    fn block_accumulator_deletion_subsumes_storage_writes() {
-        let address = Address::from([0x56; 20]);
-        let key = Word::from(1);
-        let original = AccountInfo::default().with_balance(Word::from(3));
-        let mut accumulator = BlockStateAccumulator::new();
-
-        let mut delete = account_change(address, Some(original), None);
-        delete.storage = storage_change(address, key, Word::from(5), Word::from(7), true).storage;
-        delete.visit(&mut accumulator).expect("block accumulator is infallible");
-
-        let accounts = accumulator.accounts_sorted();
-        assert_eq!(accounts.len(), 1);
-        assert_eq!(accounts[0].0, address);
-        assert!(accounts[0].1.current().is_none());
-        assert!(accumulator.storage_wipes_sorted().is_empty());
-        assert!(accumulator.storage_sorted().is_empty());
-    }
-
-    #[test]
-    fn block_accumulator_collapses_storage_wipe_write_wipe() {
-        let address = Address::from([0x52; 20]);
-        let key = Word::from(1);
-        let mut accumulator = BlockStateAccumulator::new();
-
-        let first = storage_change(address, key, Word::from(5), Word::from(7), true);
-        first.visit(&mut accumulator).expect("block accumulator is infallible");
-        storage_wipe(address).visit(&mut accumulator).expect("block accumulator is infallible");
-
-        assert!(accumulator.accounts_sorted().is_empty());
-        assert_eq!(accumulator.storage_wipes_sorted(), [address]);
-        assert!(accumulator.storage_sorted().is_empty());
-    }
-
-    #[test]
-    fn block_accumulator_keeps_account_only_and_storage_only_changes_separate() {
-        let account_address = Address::from([0x53; 20]);
-        let storage_address = Address::from([0x54; 20]);
-        let key = Word::from(1);
-        let original = AccountInfo::default().with_balance(Word::from(1));
-        let current = AccountInfo::default().with_balance(Word::from(2));
-        let mut accumulator = BlockStateAccumulator::new();
-
-        account_change(account_address, Some(original.clone()), Some(current.clone()))
-            .visit(&mut accumulator)
-            .expect("block accumulator is infallible");
-        storage_change(storage_address, key, Word::from(3), Word::from(4), false)
-            .visit(&mut accumulator)
-            .expect("block accumulator is infallible");
-
-        let accounts = accumulator.accounts_sorted();
-        assert_eq!(accounts.len(), 1);
-        assert_eq!(accounts[0].0, account_address);
-        assert_eq!(accounts[0].1.original().as_ref(), Some(&without_code(original)));
-        assert_eq!(accounts[0].1.current().as_ref(), Some(&without_code(current)));
-        assert!(accumulator.storage_wipes_sorted().is_empty());
-
-        let storage = accumulator.storage_sorted();
-        assert_eq!(storage.len(), 1);
-        assert_eq!(storage[0].0.address(), storage_address);
-        assert_eq!(storage[0].0.key(), key);
-        assert_eq!(*storage[0].1.original(), Word::from(3));
-        assert_eq!(*storage[0].1.current(), Word::from(4));
-    }
-
     #[test]
     fn eip7708_delayed_burn_logs_selfdestructs_sorted() {
         let high = Address::from([0x22; 20]);
